Split and reorganize test suite for better coverage and clarity
