+++ src/lib/mock-client-debug.test.ts
@@ -1,5 +1,9 @@
 import { DynamoDBClient, GetItemCommand } from "@aws-sdk/client-dynamodb";
-import { S3Client, GetObjectCommand } from "@aws-sdk/client-s3";
+import {
+  S3Client,
+  GetObjectCommand,
+  GetObjectCommandOutput,
+} from "@aws-sdk/client-s3";
 import {
   describe,
   test,
@@ -452,5 +456,264 @@ describe("Global Debug Configuration", () => {
       );
       expect(noMockLog).toBeDefined();
     });
+  });
+});
+
+describe("Debug Mode", () => {
+  let s3Mock: ReturnType<typeof mockClient>;
+  let s3Client: S3Client;
+  let consoleSpy: any;
+
+  beforeEach(() => {
+    s3Mock = mockClient(S3Client);
+    s3Client = new S3Client({});
+    consoleSpy = vi.spyOn(console, "log").mockImplementation(() => void 0);
+  });
+
+  afterEach(() => {
+    s3Mock.restore();
+    // eslint-disable-next-line @typescript-eslint/no-unsafe-call, @typescript-eslint/no-unsafe-member-access
+    consoleSpy.mockRestore();
+  });
+
+  test("should not log when debug is disabled", async () => {
+    s3Mock
+      .on(GetObjectCommand)
+      .resolves({ Body: "test" as unknown as GetObjectCommandOutput["Body"] });
+
+    await s3Client.send(
+      new GetObjectCommand({
+        Bucket: "test-bucket",
+        Key: "test-key",
+      }),
+    );
+
+    expect(consoleSpy).not.toHaveBeenCalled();
+  });
+
+  test("should log when debug is enabled", async () => {
+    s3Mock.enableDebug();
+    s3Mock
+      .on(GetObjectCommand)
+      .resolves({ Body: "test" as unknown as GetObjectCommandOutput["Body"] });
+
+    await s3Client.send(
+      new GetObjectCommand({
+        Bucket: "test-bucket",
+        Key: "test-key",
+      }),
+    );
+
+    // Check that configuration was logged
+    // eslint-disable-next-line @typescript-eslint/no-unsafe-assignment, @typescript-eslint/no-unsafe-call, @typescript-eslint/no-unsafe-member-access
+    const configCall = consoleSpy.mock.calls.find(
+      (call: unknown[]) =>
+        typeof call[0] === "string" &&
+        call[0].includes("Configured resolves for GetObjectCommand"),
+    );
+    expect(configCall).toBeDefined();
+
+    // Check that the command received log includes the full object
+    // eslint-disable-next-line @typescript-eslint/no-unsafe-assignment, @typescript-eslint/no-unsafe-call, @typescript-eslint/no-unsafe-member-access
+    const commandLog = consoleSpy.mock.calls.find(
+      (call: unknown[]) =>
+        typeof call[0] === "string" &&
+        call[0].includes("Received command: GetObjectCommand"),
+    );
+    expect(commandLog).toBeDefined();
+    // eslint-disable-next-line @typescript-eslint/no-unsafe-member-access
+    const commandLogText = String(commandLog?.[0] ?? "");
+    expect(commandLogText).toContain('"Bucket"');
+    expect(commandLogText).toContain('"test-bucket"');
+    expect(commandLogText).toContain('"Key"');
+    expect(commandLogText).toContain('"test-key"');
+
+    // Check that mock was found and used
+    // eslint-disable-next-line @typescript-eslint/no-unsafe-assignment, @typescript-eslint/no-unsafe-call, @typescript-eslint/no-unsafe-member-access
+    const foundCall = consoleSpy.mock.calls.find(
+      (call: unknown[]) =>
+        typeof call[0] === "string" &&
+        call[0].includes("Found 1 mock(s) for GetObjectCommand"),
+    );
+    expect(foundCall).toBeDefined();
+
+    // eslint-disable-next-line @typescript-eslint/no-unsafe-assignment, @typescript-eslint/no-unsafe-call, @typescript-eslint/no-unsafe-member-access
+    const usingCall = consoleSpy.mock.calls.find(
+      (call: unknown[]) =>
+        typeof call[0] === "string" &&
+        call[0].includes("Using mock at index 0 for GetObjectCommand"),
+    );
+    expect(usingCall).toBeDefined();
+  });
+
+  test("should log mock configuration", () => {
+    s3Mock.enableDebug();
+
+    s3Mock
+      .on(GetObjectCommand, { Bucket: "test-bucket" })
+      .resolves({ Body: "test" as unknown as GetObjectCommandOutput["Body"] });
+
+    // Check that the configuration log includes the full matcher details
+    // eslint-disable-next-line @typescript-eslint/no-unsafe-assignment, @typescript-eslint/no-unsafe-member-access
+    const firstCall = consoleSpy.mock.calls[0];
+    expect(firstCall).toBeDefined();
+    // eslint-disable-next-line @typescript-eslint/no-unsafe-member-access
+    const configLog = String(firstCall?.[0] ?? "");
+    expect(configLog).toContain("Configured resolves for GetObjectCommand");
+    expect(configLog).toContain('"Bucket"');
+    expect(configLog).toContain('"test-bucket"');
+    expect(configLog).toContain('"strict"');
+    expect(configLog).toContain("false");
+  });
+
+  test("should log resolvesOnce configuration", () => {
+    s3Mock.enableDebug();
+
+    s3Mock.on(GetObjectCommand).resolvesOnce({
+      Body: "once" as unknown as GetObjectCommandOutput["Body"],
+    });
+
+    // eslint-disable-next-line @typescript-eslint/no-unsafe-member-access
+    const call = consoleSpy.mock.calls[0]?.[0] as string;
+    expect(call).toContain("aws-sdk-vitest-mock(debug):");
+    expect(call).toContain("Configured resolvesOnce for GetObjectCommand");
+  });
+
+  test("should log rejects configuration", () => {
+    s3Mock.enableDebug();
+
+    s3Mock.on(GetObjectCommand).rejects(new Error("test error"));
+
+    // eslint-disable-next-line @typescript-eslint/no-unsafe-member-access
+    const call = consoleSpy.mock.calls[0]?.[0] as string;
+    expect(call).toContain("aws-sdk-vitest-mock(debug):");
+    expect(call).toContain("Configured rejects for GetObjectCommand");
+  });
+
+  test("should log reset operation", () => {
+    s3Mock.enableDebug();
+
+    s3Mock.reset();
+
+    // eslint-disable-next-line @typescript-eslint/no-unsafe-member-access
+    const call = consoleSpy.mock.calls[0]?.[0] as string;
+    expect(call).toContain("aws-sdk-vitest-mock(debug):");
+    expect(call).toContain("Clearing call history (mocks preserved)");
+  });
+
+  test("should log restore operation", () => {
+    s3Mock.enableDebug();
+
+    s3Mock.restore();
+
+    // eslint-disable-next-line @typescript-eslint/no-unsafe-member-access
+    const call = consoleSpy.mock.calls[0]?.[0] as string;
+    expect(call).toContain("aws-sdk-vitest-mock(debug):");
+    expect(call).toContain(
+      "Restoring original client behavior and clearing all mocks",
+    );
+  });
+
+  test("should log when no mock is found", async () => {
+    s3Mock.enableDebug();
+
+    await expect(
+      s3Client.send(
+        new GetObjectCommand({
+          Bucket: "test-bucket",
+          Key: "test-key",
+        }),
+      ),
+    ).rejects.toThrow();
+
+    expect(consoleSpy).toHaveBeenCalled();
+  });
+
+  test("should log when mock does not match", async () => {
+    s3Mock.enableDebug();
+
+    s3Mock
+      .on(GetObjectCommand, { Bucket: "other-bucket" })
+      .resolves({ Body: "test" as unknown as GetObjectCommandOutput["Body"] });
+
+    await expect(
+      s3Client.send(
+        new GetObjectCommand({
+          Bucket: "test-bucket",
+          Key: "test-key",
+        }),
+      ),
+    ).rejects.toThrow();
+
+    // Check that mocks were found
+    // eslint-disable-next-line @typescript-eslint/no-unsafe-assignment, @typescript-eslint/no-unsafe-call, @typescript-eslint/no-unsafe-member-access
+    const foundCall = consoleSpy.mock.calls.find(
+      (call: unknown[]) =>
+        typeof call[0] === "string" &&
+        call[0].includes("Found 1 mock(s) for GetObjectCommand"),
+    );
+    expect(foundCall).toBeDefined();
+
+    // Check that the no matching mock log includes the full object
+    // eslint-disable-next-line @typescript-eslint/no-unsafe-assignment, @typescript-eslint/no-unsafe-call, @typescript-eslint/no-unsafe-member-access
+    const noMatchLog = consoleSpy.mock.calls.find(
+      (call: unknown[]) =>
+        typeof call[0] === "string" &&
+        call[0].includes("No matching mock found for GetObjectCommand"),
+    );
+    expect(noMatchLog).toBeDefined();
+    // eslint-disable-next-line @typescript-eslint/no-unsafe-member-access
+    const noMatchLogText = String(noMatchLog?.[0] ?? "");
+    expect(noMatchLogText).toContain('"Bucket"');
+    expect(noMatchLogText).toContain('"test-bucket"');
+    expect(noMatchLogText).toContain('"Key"');
+    expect(noMatchLogText).toContain('"test-key"');
+  });
+
+  test("should log when one-time mock is removed", async () => {
+    s3Mock.enableDebug();
+    // eslint-disable-next-line @typescript-eslint/no-unsafe-call, @typescript-eslint/no-unsafe-member-access
+    consoleSpy.mockClear(); // Clear the configuration log
+
+    s3Mock.on(GetObjectCommand).resolvesOnce({
+      Body: "once" as unknown as GetObjectCommandOutput["Body"],
+    });
+
+    // eslint-disable-next-line @typescript-eslint/no-unsafe-call, @typescript-eslint/no-unsafe-member-access
+    consoleSpy.mockClear(); // Clear the configuration log before sending
+
+    await s3Client.send(
+      new GetObjectCommand({
+        Bucket: "test-bucket",
+        Key: "test-key",
+      }),
+    );
+
+    // Check that one-time mock removal was logged
+    // eslint-disable-next-line @typescript-eslint/no-unsafe-assignment, @typescript-eslint/no-unsafe-call, @typescript-eslint/no-unsafe-member-access
+    const removeCall = consoleSpy.mock.calls.find(
+      (call: unknown[]) =>
+        typeof call[0] === "string" &&
+        call[0].includes("Removed one-time mock for GetObjectCommand"),
+    );
+    expect(removeCall).toBeDefined();
+  });
+
+  test("should stop logging when debug is disabled", async () => {
+    s3Mock.enableDebug();
+    s3Mock.disableDebug();
+
+    s3Mock
+      .on(GetObjectCommand)
+      .resolves({ Body: "test" as unknown as GetObjectCommandOutput["Body"] });
+
+    await s3Client.send(
+      new GetObjectCommand({
+        Bucket: "test-bucket",
+        Key: "test-key",
+      }),
+    );
+
+    expect(consoleSpy).not.toHaveBeenCalled();
   });
 });
